Refs #16. refactor: moved PointCloud and McvineResolutionData to their own modules
