+++ dgsres/singlextal/plot.py
@@ -1,5 +1,7 @@
 import os, numpy as np
 from . import use_covmat
+from .pointcloud import PointCloud
+from .simdata import McvineResolutionData
 
 def createARCSAnalyticalModel(
     tau_P, tau_M, pix_r, pix_h,
@@ -29,87 +31,6 @@ def createARCSAnalyticalModel(
         sample_yml, samplethickness,
         Ei, psi_scan)
 
-
-class PointCloud:
-    
-    """resolution point cloud. can be regarded as "events" with dh,dk,dl,dE,weight data.
-    These data could be generated from mcvine simulation of the resolution function,
-    or generated from a simple normal distribution using a 4D cov matrix (see AnalyticalModel).
-    """
-    
-    def __init__(self, dhs, dks, dls, dEs, weights):
-        self.dhs = dhs
-        self.dks = dks
-        self.dls = dls
-        self.dEs = dEs
-        self.weights = weights
-        return
-    
-    def getThinSlice(
-        self, 
-        axis1=('h', -0.1,0.1, 0.002), axis2=('E', -3, 3., 0.04),
-        axis3=('k', -0.006, 0.006), axis4=('l', -0.006, 0.006)
-    ):
-        axis1_name = axis1[0]; axis1_ticks = np.arange(*axis1[1:])
-        axis2_name = axis2[0]; axis2_ticks = np.arange(*axis2[1:])
-        condition = True
-        for ax in [axis3, axis4]:
-            name = ax[0]
-            min, max = ax[1:]
-            arr = self._evts(name)
-            condition *= (arr<max) * (arr>min)
-            continue
-        Ixy, xedges, yedges = np.histogram2d(
-            self._evts(axis1_name)[condition], self._evts(axis2_name)[condition],
-            bins=(axis1_ticks, axis2_ticks), weights=self.weights[condition]) 
-        xbc = (xedges[:-1] + xedges[1:])/2
-        ybc = (yedges[:-1] + yedges[1:])/2
-        xg, yg = np.meshgrid(xbc, ybc)
-        return xg,yg,Ixy
-            
-    def _evts(self, name):
-        return getattr(self, 'd%ss' % name)
-    
-
-class McvineResolutionData:
-
-    """resolution data simulated by mcvine. 
-    The data was simulated earlier and saved to disk. 
-    This class handles locating the data files and reading the data, 
-    and convert the data to different forms: point cloud, cov matrix, etc.
-    """
-    
-    def __init__(self, parent_dir, dirname_template='E%s_hkl%s'):
-        self.parent_dir = parent_dir
-        self.dirname_template = dirname_template
-        return
-    
-    def path(self, hkl, E):
-        return os.path.join(self.parent_dir, self.dirname_template % (E, '%s,%s,%s' % tuple(hkl)))
-    
-    def loadData(self, hkl, E):
-        p = self.path(hkl, E)
-        return self._loadData(p)
-    
-    def loadPointCloud(self, hkl, E):
-        dhs, dks, dls, dEs, probs = self.loadData(hkl, E)
-        return PointCloud(dhs, dks, dls, dEs, probs)
-    
-    def _loadData(self, outdir1):
-        dhkls = np.load('%s/dhkls.npy' % outdir1)
-        dEs = np.load('%s/dEs.npy' % outdir1)
-        probs = np.load('%s/probs.npy' % outdir1)
-        dhs,dks,dls = dhkls.T
-        # there might be unreasonable data points with unreasonable weights
-        mask = (dhs> -2.)*(dhs<2.) \
-            * (dks> -2.)*(dks<2.) \
-            * (dls> -2.)*(dls<2.) 
-        return np.array([dhs[mask], dks[mask], dls[mask], dEs[mask], probs[mask]])
-    
-    def computeCovMat(self, hkl, E):
-        data = self.loadData(hkl, E)
-        Data = data[:4]; probs = data[-1]
-        return np.cov(Data, aweights=probs)
 
 class AnalyticalModel:
